+++ main.py
@@ -1,17 +1,11 @@
-import importlib
-import os
-
-import config
 from LSL import LSL
-#import sys
-#sys.path.append('/Documents/DataCollector')
 from tests.TestGUI import TestGUI
 from tests.Action import Action
 
 
 class DataCollectorApp:
     """
-    A collection of functions to setup the Tkinter GUI
+    A collection of functions to initialize the GUI, tests, and setup test logic.
     """
 
     @staticmethod
@@ -21,50 +15,48 @@ class DataCollectorApp:
 
         :param test_name: Name of the test being run
         """
-
-        test_arry = ['Blink', 'Float left']
-
-        """ Trying another way so we dont have 20 test files """
-        test = Action(test_name)
+        test = Action(test_name, 0)
         test.start()
 
+        def callback(test_complete):
+            if DataCollectorApp.__all_tests_complete():
+                TestGUI.control_window.quit()
+                print("All tests complete.")
 
-        # Dynamically import the test from tests package & construct it w/ no parameters
-        # test_class = getattr(importlib.import_module(f"tests.{test_name}"), test_name)
-        # test = test_class()
+            if not test_complete:  # If test is not complete
+                TestGUI.tests[test_name]["trial_number"] += 1  # Increase trial number OUTSIDE OF THREAD!!!
 
-        # test.start()  # Start test thread
+        test.set_callback(callback)
 
     @staticmethod
     def run():
         """
         Main function for adding buttons that run tests to the GUI and initializing the LSL streams & GUI.
         """
-        # Create tests directory TODO make for each subject
-        os.makedirs(config.DATA_PATH, exist_ok=True)
-
         # Initialize streams & GUI
-        TestGUI.init_gui()
         LSL.init_lsl_stream()
+        TestGUI.init_gui()
 
-        # Locate all tests (filename should be the same as test name)
-        '''
-        test_directory = 'tests'
-        test_names = [filename.split('.')[0]
-                      for filename in os.listdir(test_directory)
-                      if filename.endswith('.py') and not filename.startswith('Test') and not filename.startswith('__init__')]
-        '''
+        test_names = ['Blink', 'Brow Furrow', 'Brow Unfurrow', 'Stationary Floating', 'Float Left', 'Float Right',
+                      'Float Up', 'Float Down', 'Select', 'Eyes Open', 'Eyes Closed']
 
-
-        test_names = ['Blink', 'Brow Furrow', 'Brow Unfurrow', 'Stationary Floating', 'Float Left', 'Float Right', 'Float Up', 'Float Down', 'Select', 'Eyes Open', 'Eyes Closed']
-
-        # Add each test button to the GUI that calls the run_test method above w/ the test name
         for test_name in test_names:
+            # Add button to test
             TestGUI.add_button(test_name, lambda name=test_name: DataCollectorApp.run_test(name))
 
-        # TODO how do we know if all tests are complete? Close window when done.
-
         TestGUI.control_window.mainloop()
+
+    @staticmethod
+    def __all_tests_complete() -> bool:
+        """
+        Helper function to check the current state of all tests in the GUI.
+
+        :return: True if all tests are complete, false otherwise.
+        """
+        for test in TestGUI.tests.keys():
+            if not TestGUI.tests[test]['completed']:
+                return False
+        return True
 
 
 if __name__ == '__main__':