+++ tests/TestGUI.py
@@ -1,4 +1,4 @@
-import time
+import os.path
 import tkinter as tk
 
 import matplotlib.pyplot as plt
@@ -16,8 +16,9 @@ class TestGUI:
     control_window = None
     display_window = None
 
-    test_buttons = {}
-    test_status = {}
+    # Setup test states: A dictionary with test name keys corresponding to sub-dictionaries with lambda, button,
+    # trial_number, and completed parameters
+    tests = {}
     current_test = None
 
     subject_number = ""
@@ -41,33 +42,46 @@ class TestGUI:
     def add_button(test_name, test_lambda):
         """
         Adds a button to the test window with its name and function to run.
+
+        :param test_name: The name of the test that the button will be assigned to.
+        :param test_lambda: The function that the test will be ran with.
         """
+        # COnfigure button
         btn = tk.Button(TestGUI.control_window, text=test_name)
         btn.config(command=test_lambda, bg='red')
         btn.pack()
-        TestGUI.test_buttons[test_name] = btn
-        TestGUI.test_status[test_name] = False
+
+        # Configure test state
+        TestGUI.tests[test_name] = { 'lambda': test_lambda, 'button': btn, 'trial': 0, 'completed': False}
         print("Added test: " + test_name)
 
     @staticmethod
-    def confirm_test():
+    def confirm_current_test(test_data_path: str) -> bool:
         """
-        Run the test and prompt the user to confirm or deny the data
-        """
-        TestGUI.__show_data_and_confirm()
+        Run the test and prompt the user to confirm or deny the data.
 
-        # TODO if denied, move to next test and allow for a rerun
+        :param test_data_path: The path to the current test data FOLDER.
+        """
+        # Confirm data
+        TestGUI.__show_data_and_confirm(os.path.join(test_data_path, config.FILENAME))
+
+        # Log finalized test status
+        print(f"{TestGUI.current_test} - Trial {TestGUI.tests[TestGUI.current_test]['trial_number']}: {'Complete' if TestGUI.tests[TestGUI.current_test]['completed'] else 'Discarded'}")
+
+        return TestGUI.tests[TestGUI.current_test]['completed']
 
     @staticmethod
     def disable_buttons(test_name):
         """
         A function to disable buttons while a test is running.
+
+        :param test_name: The name of the current test running
         """
-        for button in TestGUI.test_buttons.values():
-            button.config(state="disabled")
+        for test in TestGUI.tests.keys():
+            TestGUI.tests[test]['button'].config(state="disabled")
 
         TestGUI.current_test = test_name
-        TestGUI.test_buttons[TestGUI.current_test].config(bg="yellow")
+        TestGUI.tests[TestGUI.current_test]['button'].config(bg="yellow")
 
     #
     # HELPER METHODS
@@ -79,22 +93,21 @@ class TestGUI:
         A function to enable buttons after a test has been completed.
         """
         # Reset the buttons
-        for test in TestGUI.test_buttons.keys():
-            if not TestGUI.test_status[test]:  # If test is not complete, re-enable button
-                TestGUI.test_buttons[test].config(state="normal")
-
-        TestGUI.current_test = None
+        for test in TestGUI.tests.keys():
+            if not TestGUI.tests[test]['completed']:  # If test is not complete, re-enable button
+                TestGUI.tests[test]['button'].config(state="normal")
 
     @staticmethod
-    def __show_data_and_confirm():
+    def __show_data_and_confirm(test_data_path: str):
         """
         Popup the data confirmation window and check if it should be accepted or rejected.
-        TODO return boolean to determine whether to move onto a new test.
+
+        :param test_data_path: The path to the test data CSV.
         """
         # Setup the window and canvas
         popup = tk.Toplevel()
         popup.wm_title("Data Confirmation")
-        popup.attributes('-zoomed', True)
+        popup.state('zoomed')
 
         canvas = tk.Canvas(popup)
         scrollbar = tk.Scrollbar(popup, orient="vertical", command=canvas.yview)
@@ -113,7 +126,7 @@ class TestGUI:
         # Make it scrollable so we don't have to click the tiny scrollbar
         TestGUI.__enable_scroll(canvas)
 
-        data_df = pd.read_csv(config.COLLECTED_DATA_PATH)
+        data_df = pd.read_csv(test_data_path)
 
         # Figure out how many graphs we can fit on the screen
         screen_width = popup.winfo_screenwidth()
@@ -121,8 +134,6 @@ class TestGUI:
 
         if graphs_per_row == 0:
             graphs_per_row = 1
-
-        # TODO do we want to draw the graphs from the timestamp of the past test?
 
         # Loop through each EEG and draw the graphs
         idx = 0
@@ -154,7 +165,10 @@ class TestGUI:
 
         # TODO explicitly destroy the graphs
 
-        popup.mainloop()
+        # popup.mainloop()
+        # Force popup to be on top & halt program execution
+        popup.grab_set()
+        TestGUI.control_window.wait_window(popup)
 
     @staticmethod
     def __confirm_data(popup):
@@ -166,13 +180,11 @@ class TestGUI:
         popup.destroy()
 
         # Set button color to green and disable to mark that the test has been completed
-        TestGUI.test_buttons[TestGUI.current_test].config(bg="green")
-        TestGUI.test_buttons[TestGUI.current_test].config(state="disabled")
-        TestGUI.test_status[TestGUI.current_test] = True
+        TestGUI.tests[TestGUI.current_test]['button'].config(bg="green")
+        TestGUI.tests[TestGUI.current_test]['button'].config(state="disabled")
+        TestGUI.tests[TestGUI.current_test]['completed'] = True
 
         TestGUI.__enable_buttons()
-
-        # TODO save test and mark as done
 
     @staticmethod
     def __deny_data(popup):
@@ -183,10 +195,8 @@ class TestGUI:
         """
         # Close window and then clear all the data
         popup.destroy()
-        TestGUI.test_buttons[TestGUI.current_test].config(bg="red")
+        TestGUI.tests[TestGUI.current_test]['button'].config(bg="red")
         TestGUI.__enable_buttons()
-
-        # TODO if this is called, move to next test
 
     @staticmethod
     def __prompt_subject_number():
@@ -213,7 +223,9 @@ class TestGUI:
         submit_button = tk.Button(popup, text='Begin', command=submit)
         submit_button.pack()
 
+        # Force popup to be on top & halt program execution
         popup.grab_set()
+        TestGUI.control_window.wait_window(popup)
 
     @staticmethod
     def __enable_scroll(canvas):